moved clone to actions, metric table hidden by default
