+++ hawc/apps/study/templates/study/study_clone.html
@@ -9,7 +9,7 @@
        hx-indicator="#spinner-{{object.pk}}"
        hx-target="#src-assessment-details"
        hx-encoding="multipart/form-data">
-      Fetch Available Studies
+      Fetch studies
     </a>
     <i class="fa fa-spinner fa-spin htmx-indicator ml-3" id="spinner-{{object.pk}}" aria-hidden="true"></i>
     <br><br>
